requirements.txt
Flask
Flask-Cors
python-dotenv
gunicorn
requests
yfinance